Move city_furniture admin utils to a separate file
